requirements.txt
aws_cdk.core==1.117.0
aws_cdk.aws-ec2==1.117.0